extract CustomGraphPass
